signimage: add older shell scripts used to generate
the key database

These scripts will serve as templates for new versions, which will
be designed to be used by YourFritz users themselves.

The whole database is maintained as a signed XML document, the
YourFritz key is used to sign it.

To handle/manipulate the database, some XML tools are needed.

- xmllint
- xmlsec1
- xsltproc